P2-12b: share renderStat + HelpHint across SummaryCards
Extract two pieces that were copy-pasted across the entity SummaryCards,
with zero change to rendered output:
- renderStat() -> src/lib/utils.ts (was duplicated in region/ao/area/sector),
  now unit-tested.
- HelpHint -> the mobile-aware Tooltip/Popover "?" affordance (was an inline
  isMobile hook + ~20-line block in region/area/sector).

region/ao/area/sector SummaryCards now import both; the per-row markup is
intentionally left inline (region uses text-sm, area/sector don't — unifying
that StatRow needs a visual check). pax SummaryCard is structurally different
(11 rows, no renderStat) and deferred.
